fine tune some CI test zip linux m0hamed_rootkit
